Rework query GetData (SyncWithBatch)
Previously, query GetData could synchronize improperly if you had
a command list with Begin/End in it.

Rather than having to explicitly point to a batch to declare a
dependency, use fence-like IDs instead. When executing a
command list, advance an entire "generation" of IDs. This acts
as a signal to users that would need to synchronize against batch
execution that there could be untracked commands, and to assume
the worst.