Remove DXBC Signer dependency from the translation layer
Signing can be done as a post process by the projects that need it. For
example, 9on12 will take a dependency on DXBC Signer and sign the blob
after calling CDXBCBuilder::GetFinalDXBC.

Laving the Dxbc Parser as a separate lib since it is cleaner to have the
signer depend only on that than the entire translation layer lib.