Rewrite residency library
Rather than bolting on the D3DX12 residency library, trim the fat
and integrate it a bit more into the translation layer. Notably,
we remove sync points and just use the existing 3-queue-fences,
which removes the need to store fences in private data, as well as
the need to signal a second fence on every submission. It also removes
a copy of the residency set on every submit (since this codebase only
submits on command list at a time anyway).

To resolve include order issues, the implementation of a few functions
is moved to a source file. Further changes will follow.