Update setting of dirty bits to take root signature SRV/CB/Sampler bucket sizes into account
This make the setting of StatesToReassert bits more correct, which fixes
an issue where the Apply*Helper functions were trying to bind invalid
descriptor tables with bad underlying data. While this didn't seem to be
affecting applications, it did cause a breaking error to be logged in
the debug layer.